refactor(deps): move DurationFormat abstract operations to ecma402-abstract
Move 5 DurationFormat abstract operations to ecma402-abstract/DurationFormat/:
- DurationRecordSign, GetDurationUnitOptions, IsValidDurationRecord
- ToDurationRecord, ToIntegerIfIntegral

PartitionDurationFormatPattern stays in intl-durationformat (imports
polyfill-specific DurationFormat class and internal slots).

Also:
- Create ecma402-abstract/types/duration.ts with DurationRecord and DurationInput
- Move TABLE_1 constant to DurationFormat/constants.ts
- Re-export types from intl-durationformat/src/types.ts for backward compat
