+++ packages/ecma402-abstract/DurationFormat/ToIntegerIfIntegral.ts
