+++ packages/ecma402-abstract/DurationFormat/constants.ts
@@ -1,0 +1,14 @@
+import type {DurationRecord} from '#packages/ecma402-abstract/types/duration.js'
+
+export const TABLE_1: Array<keyof DurationRecord> = [
+  'years',
+  'months',
+  'weeks',
+  'days',
+  'hours',
+  'minutes',
+  'seconds',
+  'milliseconds',
+  'microseconds',
+  'nanoseconds',
+]
