+++ packages/ecma402-abstract/types/duration.ts
@@ -1,0 +1,25 @@
+export interface DurationInput {
+  years?: number
+  months?: number
+  weeks?: number
+  days?: number
+  hours?: number
+  minutes?: number
+  seconds?: number
+  milliseconds?: number
+  microseconds?: number
+  nanoseconds?: number
+}
+
+export interface DurationRecord {
+  years: number
+  months: number
+  weeks: number
+  days: number
+  hours: number
+  minutes: number
+  seconds: number
+  milliseconds: number
+  microseconds: number
+  nanoseconds: number
+}
