+++ packages/intl-durationformat/src/core.ts
@@ -8,9 +8,9 @@ import {GetOption} from '#packages/ecma402-abstract/GetOption.js'
 import {SupportedLocales} from '#packages/ecma402-abstract/SupportedLocales.js'
 import {invariant} from '#packages/ecma402-abstract/utils.js'
 import {ResolveLocale} from '@formatjs/intl-localematcher'
-import {GetDurationUnitOptions} from './abstract/GetDurationUnitOptions.js'
+import {GetDurationUnitOptions} from '#packages/ecma402-abstract/DurationFormat/GetDurationUnitOptions.js'
 import {PartitionDurationFormatPattern} from './abstract/PartitionDurationFormatPattern.js'
-import {ToDurationRecord} from './abstract/ToDurationRecord.js'
+import {ToDurationRecord} from '#packages/ecma402-abstract/DurationFormat/ToDurationRecord.js'
 import {getInternalSlots} from './get_internal_slots.js'
 import {numberingSystemNames} from './numbering-systems.generated.js'
 import {TIME_SEPARATORS} from './time-separators.generated.js'
